[cairo] move linemetrics to own module
